+++ packages/popmotion/src/action/types.ts
@@ -1,12 +1,14 @@
 import { IObserver, ObserverProps } from '../observer/types';
 
-export type ActionInit = (observer: IObserver) => ColdSubscription | void;
+export type ActionInit = (
+  observer: IObserver
+) => Partial<ColdSubscription> | void;
 
 export interface ColdSubscription {
-  stop?: () => void;
+  stop: () => void;
   [key: string]: Function | undefined;
 }
 
 export type ActionProps = ObserverProps & {
-  init: ActionInit
+  init: ActionInit;
 };