+++ packages/react-pose/src/_tests/index.test.tsx
@@ -1,9 +1,9 @@
-import React from 'react';
+import React, { forwardRef } from 'react';
 import posed from '../posed';
 import PoseGroup from '../components/Transition/PoseGroup';
-import { render, cleanup } from 'react-testing-library'
+import { render, cleanup } from 'react-testing-library';
 
-afterEach(cleanup)
+afterEach(cleanup);
 
 const Parent = posed.div({
   init: { x: 10, transition: { duration: 30 } },
@@ -17,7 +17,7 @@ const Parent = posed.div({
   dynamicExit: {
     x: ({ x }) => x,
     transition: { duration: 30 }
-  },
+  }
 });
 
 const Child = posed.div({
@@ -33,7 +33,7 @@ const Child = posed.div({
   dynamicExitDuration: {
     y: 85,
     transition: ({ i }) => ({ duration: (i + 1) * 30 })
-  },
+  }
 });
 
 test('posed: initial state - named initial pose', () => {
@@ -62,6 +62,34 @@ test('posed: initial state - auto `init` pose', () => {
 
   expect(x).toBe(10);
   expect(y).toBe(15);
+});
+
+test('posed: nested poses get merged together', () => {
+  let parentX = 0;
+  let childX = 0;
+  let childY = 0;
+
+  const Parent = posed.div({
+    foo: { x: 20, transition: { duration: 30 } }
+  });
+
+  const Child = posed.div({
+    foo: { x: 10, transition: { duration: 30 } },
+    bar: { y: 30, transition: { duration: 30 } }
+  });
+
+  render(
+    <Parent pose="foo" onValueChange={{ x: v => (parentX = v) }}>
+      <Child
+        pose="bar"
+        onValueChange={{ x: v => (childX = v), y: v => (childY = v) }}
+      />
+    </Parent>
+  );
+
+  expect(parentX).toBe(20);
+  expect(childX).toBe(10);
+  expect(childY).toBe(30);
 });
 
 test('posed: mount animation with `initialPose`', () => {
@@ -227,37 +255,31 @@ test('PoseGroup: onRest fires when exit pose starts during exit pose', () => {
   return new Promise(resolve => {
     class Group extends React.Component {
       state = {
-        list: range(6),
-      }
+        list: range(6)
+      };
 
       componentDidMount() {
         this.pop2();
       }
 
       pop2 = () => {
-        const { list } = this.state
+        const { list } = this.state;
 
         if (!list.length) {
           return;
         }
 
         this.setState({ list: list.slice(0, -2) });
-      }
+      };
 
       render() {
         return (
-          <PoseGroup
-            exitPose="dynamicExitDuration"
-            onRest={resolve}
-          >
-            {this.state.list.map(i =>
-              <Child
-                i={i}
-                key={i}
-                onPoseComplete={this.pop2}
-              />)}
+          <PoseGroup exitPose="dynamicExitDuration" onRest={resolve}>
+            {this.state.list.map(i => (
+              <Child i={i} key={i} onPoseComplete={this.pop2} />
+            ))}
           </PoseGroup>
-        )
+        );
       }
     }
 
@@ -390,25 +412,21 @@ test('PoseGroup: Provides group props to children on mount', () => {
   let x = 0;
   let y = 0;
 
-  const ChildWithGroupProps = posed(({ innerRef, hostRef, text, ...props }) => (
-    <div ref={innerRef} {...props}>{`text: ${text}`}</div>
-  ))({});
+  const ChildWithGroupProps = posed(
+    forwardRef(({ text, ...props }, innerRef) => (
+      <div ref={innerRef} {...props}>{`text: ${text}`}</div>
+    ))
+  )({});
 
   const Group = () => (
-    <PoseGroup
-      animateOnMount
-      text="foo bar"
-    >
-      <ChildWithGroupProps
-        data-testid="child"
-        key="a"
-      />
+    <PoseGroup animateOnMount text="foo bar">
+      <ChildWithGroupProps data-testid="child" key="a" />
     </PoseGroup>
   );
 
   const { getByTestId, debug } = render(<Group />);
-  const child = getByTestId('child')
-  expect(child.textContent).toBe('text: foo bar')
+  const child = getByTestId('child');
+  expect(child.textContent).toBe('text: foo bar');
 });
 
 test('PoseGroup: `onPoseComplete` gets called for leaving child', () => {
@@ -421,7 +439,7 @@ test('PoseGroup: `onPoseComplete` gets called for leaving child', () => {
             onPoseComplete={pose => {
               if (pose === 'enter') {
                 rerender(<Group isVisible={false} />);
-                return
+                return;
               }
 
               expect(pose).toBe('exit');
@@ -469,5 +487,43 @@ test('PoseGroup: special pre-enter pose', () => {
     rerender(<Group isVisible />);
     expect(x).toBe(20);
     expect(y).toBe(25);
+  });
+});
+
+test('StrictMode: PoseGroup removes children correctly', () => {
+  return new Promise(resolve => {
+    const First = posed.div({
+      enter: { opacity: 1 },
+      exit: { opacity: 0 }
+    });
+
+    const Second = posed.div({
+      enter: { opacity: 1 },
+      exit: { opacity: 0 }
+    });
+
+    const Group = props => (
+      <React.StrictMode>
+        <PoseGroup {...props} />
+      </React.StrictMode>
+    );
+
+    const { rerender } = render(
+      <Group>
+        <First
+          key="first"
+          onPoseComplete={pose => {
+            expect(pose).toBe('exit');
+            resolve();
+          }}
+        />
+      </Group>
+    );
+
+    rerender(
+      <Group>
+        <Second key="second" />
+      </Group>
+    );
   });
 });