+++ packages/react-pose/src/posed.tsx
@@ -1,5 +1,4 @@
-import * as React from 'react';
-import { ReactElement } from 'react';
+import React, { forwardRef, ComponentType, HTMLProps } from 'react';
 import { PoseElement, PoseParentConsumer } from './components/PoseElement';
 import supportedElements from './utils/supported-elements';
 import {
@@ -7,6 +6,7 @@ import {
   PoseContextProps
 } from './components/PoseElement/types';
 import { DomPopmotionConfig } from 'popmotion-pose';
+import { warning } from 'hey-listen';
 
 type DomPopmotionConfigFactory<T> = (
   props: PoseElementProps & T
@@ -14,36 +14,45 @@ type DomPopmotionConfigFactory<T> = (
 
 export type ComponentFactory<T> = (
   poseConfig?: DomPopmotionConfig | DomPopmotionConfigFactory<T>
-) => (props: PoseElementProps & T) => ReactElement<T>;
+) => ComponentType<any>;
 
 export type Posed = {
-  <T>(component: React.ComponentType<T>): ComponentFactory<T>;
-  [key: string]: ComponentFactory<React.HTMLProps<any>>;
+  <T>(component: ComponentType<T>): ComponentFactory<T>;
+  [key: string]: ComponentFactory<HTMLProps<any>>;
 };
 
-const componentCache = new Map<
-  string | React.ComponentType,
-  ComponentFactory<any>
->();
+const componentCache = new Map<string | ComponentType, ComponentFactory<any>>();
 
-const createComponentFactory = (key: string | React.ComponentType) => {
+const createComponentFactory = (key: string | ComponentType) => {
   const componentFactory: ComponentFactory<any> = (poseConfig = {}) => {
-    return ({ withParent = true, ...props }) => {
+    // TODO: Replace functional context with new class property API
+    return forwardRef(({ withParent = true, ...props }, ref) => {
+      warning(
+        props.innerRef === undefined,
+        'innerRef is deprecated. Please use ref instead.'
+      );
+
       return !withParent || props.parentValues ? (
-        <PoseElement poseConfig={poseConfig} elementType={key} {...props} />
+        <PoseElement
+          poseConfig={poseConfig}
+          innerRef={ref}
+          elementType={key}
+          {...props}
+        />
       ) : (
         <PoseParentConsumer>
           {(parentCtx: PoseContextProps) => (
             <PoseElement
               poseConfig={poseConfig}
               elementType={key}
+              innerRef={ref}
               {...props}
               {...parentCtx}
             />
           )}
         </PoseParentConsumer>
       );
-    };
+    });
   };
 
   componentCache.set(key, componentFactory);
@@ -51,12 +60,12 @@ const createComponentFactory = (key: string | React.ComponentType) => {
   return componentFactory;
 };
 
-const getComponentFactory = (key: string | React.ComponentType) =>
+const getComponentFactory = (key: string | ComponentType) =>
   componentCache.has(key)
     ? componentCache.get(key)
     : createComponentFactory(key);
 
-const posed = ((component: React.ComponentType | string) =>
+const posed = ((component: ComponentType | string) =>
   getComponentFactory(component)) as Posed;
 
 supportedElements.reduce((acc, key) => {