+++ packages/stylefire/src/css/value-types.ts
@@ -10,7 +10,7 @@ import {
 
 const valueTypes: { [key: string]: ValueType } = {
   // Color props
-  color: color,
+  color,
   backgroundColor: color,
   outlineColor: color,
   fill: color,
@@ -22,24 +22,45 @@ const valueTypes: { [key: string]: ValueType } = {
   borderRightColor: color,
   borderBottomColor: color,
   borderLeftColor: color,
+  borderWidth: px,
+  borderTopWidth: px,
+  borderRightWidth: px,
+  borderBottomWidth: px,
+  borderLeftWidth: px,
   borderRadius: px,
+  borderTopLeftRadius: px,
+  borderTopRightRadius: px,
+  borderBottomRightRadius: px,
+  borderBottomLeftRadius: px,
 
-  // Positioning
+  // Positioning props
   width: px,
   maxWidth: px,
   height: px,
   maxHeight: px,
   top: px,
-  left: px,
-  bottom: px,
   right: px,
+  bottom: px,
+  left: px,
 
-  // Transform properties
+  // Spacing props
+  padding: px,
+  paddingTop: px,
+  paddingRight: px,
+  paddingBottom: px,
+  paddingLeft: px,
+  margin: px,
+  marginTop: px,
+  marginRight: px,
+  marginBottom: px,
+  marginLeft: px,
+
+  // Transform props
   rotate: degrees,
   rotateX: degrees,
   rotateY: degrees,
   rotateZ: degrees,
-  scale: scale,
+  scale,
   scaleX: scale,
   scaleY: scale,
   scaleZ: scale,