+++ playground/index.js
@@ -29,7 +29,8 @@ import {
   DragChildren,
   SingleFilter,
   DynamicPositionedPose,
-  TransitionStutter
+  TransitionStutter,
+  Refs
 } from './plugins/react-pose';
 
 storiesOf('Plugin: Pose', module)
@@ -51,6 +52,7 @@ storiesOf('Plugin: React Pose', module)
   .add('group', () => <Group />)
   .add('email list', () => <EmailList />)
   .add('user-defined poses within PoseGroup', () => <UserPoseInPoseGroup />)
+  .add('refs', () => <Refs />)
   .add('pose key test', () => <PoseKey />)
   .add('hover', () => (
     <Hover onPressStart={console.log} onPressEnd={console.log} />
@@ -67,9 +69,11 @@ storiesOf('Plugin: React Pose', module)
   .add('dynamic positioned pose', () => <DynamicPositionedPose />)
   .add('transition stuttering', () => <TransitionStutter />);
 
-import { SVG } from './stylefire';
+import { SVG, CSS } from './stylefire';
 
-storiesOf('stylefire', module).add('svg transform', () => <SVG />);
+storiesOf('stylefire', module)
+  .add('svg transform', () => <SVG />)
+  .add('css', () => <CSS />);
 
 // import { Decay, DecayModifyTarget, DecayPower, DecayTimeConstant } from './decay';
 // import { Keyframes, KeyframesDuration, KeyframesLoop, KeyframesLinear, KeyframesTimes } from './keyframes';