+++ packages/popcorn/CHANGELOG.md
@@ -2,6 +2,46 @@
 
 Popcorn adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0] 2018-09-30
+## [0.3.2] 2018-11-28
+
+### Fixing
+
+- Converting `interpolate` to generic types.
+
+## [0.3.1] 2018-11-28
+
+### Fixing
+
+- Explicit types for `interpolate`.
+
+## [0.3.0] 2018-11-27
+
+### Adding
+
+- Support for `clamp` and `ease` options in `interpolate`.
+
+## [0.2.0] 2018-11-16
+
+### Removed
+
+- `dilate` [#612](https://github.com/Popmotion/popmotion/pull/612)
+
+## [0.1.2] 2018-11-05
+
+### Fixed
+
+- Fixed Typescript import paths [#607](https://github.com/Popmotion/popmotion/issues/607)
+
+## [0.1.1] 2018-11-01
+
+### Republish
+
+## [0.1.0] 2018-10-31
+
+### Added
+
+- Exposing `springForce, springForceLinear, springForceExpo` - these are currently private for use in Popmotion libraries.
+
+## [0.0.1] 2018-09-30
 
 ### First publish